refactor(wallet-cli): extract ensureOwnerOnlyDirectory and forward resolved logger
Address PR review feedback:

- Extract the daemon data-dir mkdir + 0o700 chmod into a dedicated
  `ensureOwnerOnlyDirectory` in `data-dir.ts`, moving the access-control
  rationale into its JSDoc. This gives the owner-only guarantee a named
  import boundary the daemon-entry test can assert against directly.
- Forward the resolved `logFn` (not the raw optional `log`) to
  `subscribeToChanges` in `createWallet`, so persistence-write failures
  always use the daemon's logger instead of falling through to the
  console fallback inconsistently.
